New Prefab: EventSystemList (+ folder change)
I coded EventSystemList.cs for the respective EventSystem Prefab.

SYSTEM (list):
Put any Unity Event into Dialogue Events () - triggering animation, sound, method, ...
Enter- and exitDelay (in seconds) will say how long the script waits before invoking the events (enterDelay) or moving to the next Element in the list (exitDelay).
Name is just for organisation. Please fill it always for a better overview and debugging.
CurrentEvent is the index for the list. (-1 is before the start)

TOOLS for easier Design:
This are conditions when to start the system.
TriggerOnStart: self-explaining
TriggerOnPlayerEnter: If object with tag "Player" enters the 2D trigger of the EventSystemList object
TriggerOnEvent: Can be triggered by "StartSystem()" in a unity event or code
TriggerOnCustomTagEnter & CustomTag: If an object with the CustomTag enters the trigger-collider (as custom Tag is a string, make sure that the customTag is written correctly!)

Enable Rest: If true, you can reset the system to -1 by "ResetSystem()"

__________________________________________

Moreover, I created the folder DialogueEventSystem in Prefabs to put Ahmet's DialogueEventSystem and my DialogueEventSystem.prefab in ther