Prefab for lighter activation
This prefab can be attached to the player to play the lighter effect. Hold mouse-wheel and move the wheel down (scroll) to activate it.
I wanted to test the way of input in this way.

inputDelayThreshold = delay between pressing wheel and scrolling down
wheelThreshold = How much needs the wheel to be scrolled down, 0 or higher
lightIntensityDivider = How fast does the light turn off
flameLight = reference to Light
flameParticle = reference to Flame