Simple Item system + burning paper works now
ItemBehaviour.cs (attached to player):

For the items, there are 3 versions:
AtStart: Item where it is at the beginning
AtPlayer: Item holded by player
AtEnd: Item where it is at the end
For this you have to create 3 different gameObjects and place them in the level

For Debug only:
O + E : Take item
O + R : If you hold item, leave item at the setPosition
O + Q: If you hold item, drop it at the position you're holding it

For the game, you can call some functions to take, leave or drop something. You won't be able to pick sth up again

I already connected this to the (burning) paper/file. You can take it by debug mode, and then activate the lighter outside of the house. This will directly drop and burn the papers.
_______

Moreover I created a small script for the burning paper to ensure that the player can move through it after it fell. (It needs a collider tho so that it does not fall thorugh the ground - the ground needs the "Ground" Tag btw now)
_______

Some small fix at Ahmet's Player Movement Script: Input gets set to 0 for movement lock