TransitionScreens Prefab
Add elements in the list and attach images in the scene to it. The images will be at Start invisible and then fade in after each other.

Some remarks: "timeInTotal" will always calculate the total time, you can only cange it by changing the fadeInTime, fadeOutTime and timeBetweenFades

The event will be triggered at the start of the fadeIn.

earlierTransitionInSec allows an earlier fadeIn of the next image, but only after the image before has started to fade out.