LighteractivationScript: variables public get
onFire & isEmmittingSpark can be read from other scripts

isEmmitingSpark will be set to true if the mouse wheel is pressed /will be set to false if the lighter is properly activated or the inputDelay is 0 again (a short time after pressing the mouse wheel).

onFire is true while the lighter is properly activated.

(While both variables are true, there will be light in the scene.)