LukasNarrativeTesting : fog + mouse sensivity adjusted
fog is now in a seperate section of the hierachy and contains several layers of fog mostly overlaying the sound orbs. Moreover, there are small particles, but you can barely see them

Furthermore, I changed the mouse sensitvity curve for connected orbs (it's easier now to drag) in this scene.
I will push the change in the prefab as soon as I know that Natalie has no unpushed changes on the prefab.