Animation Laying lamp down working + other
Code for the animation: Attach LayingDown.cs to an object with a collider. If the player clicks on the collider and holds, then the first part of the anim is played. After holding longer, the  second part is played as well.
Made some changes for the anim tree together with Kassandra.

The code / tree structure is still a bit messy as I have problems to understand the  good usage of the tree's conditions in code. I will try to clear it, as well take care of level progression